pyproject.toml
[project]
name = "python-hon-usbhid"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "argparse>=1.4.0",
    "hidapi>=0.14.0.post4",
]

[tool.pylint.'MESSAGES CONTROL']
max-line-length = 120
disable = """
    too-many-arguments,
    line-too-long,
    c-extension-no-member,
"""


